refactor: import TypeWithID from payload/shared in internal packages
Follow-up to removing TypeWithID from the public barrel. Re-expose it
under payload/shared and update all internal packages that were
importing TypeWithID from 'payload' to use 'payload/shared'.
